tools/requirements.txt
requests>=2.31.0
PyYAML>=6.0.2
beautifulsoup4>=4.12.3
lxml>=5.2.1
python-dateutil>=2.9.0
